many many casings

+++ src/main/java/gregtech/api/multitileentity/enums/GT_MultiTileCasing.java
@@ -8,13 +8,13 @@ public enum GT_MultiTileCasing {
     Chemical(1),
     NONE(GT_Values.W);
 
-    private final short meta;
+    private final int meta;
 
     GT_MultiTileCasing(int meta) {
-        this.meta = (short) meta;
+        this.meta = meta;
     }
 
-    public short getId() {
+    public int getId() {
         return meta;
     }
 }
