+++ src/main/java/gregtech/common/tileentities/machines/multiblock/Macerator.java
@@ -84,7 +84,7 @@ public class Macerator extends StackableController<Macerator> {
     }
 
     @Override
-    public short getCasingMeta() {
+    public int getCasingMeta() {
         return 18000;
     }
